test_requirements.txt
coverage==4.2
Django==1.8.4
djangorestframework==3.3.2
pytest==3.0.5
pytest-cov==2.4.0
pytest-django==3.1.2
